requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.6.1
PyYAML==6.0.2
httpx==0.27.2
tenacity==9.0.0
python-multipart==0.0.12
jinja2==3.1.4

# LangGraph / LangChain
langgraph==0.2.62
# NOTE: langgraph pins a compatible langchain-core;
# avoid hard pin here to prevent resolver conflicts.

# Packet parsing
scapy==2.5.0
dpkt==1.9.8

# ML
numpy==2.1.3
scikit-learn==1.5.2
joblib==1.4.2

# Storage
SQLAlchemy==2.0.36
aiosqlite==0.20.0
asyncpg==0.30.0

# Observability / alerts
python-json-logger==3.2.1

# Tests
pytest==8.3.3
pytest-asyncio==0.24.0

